+++ src/main/java/faang/school/achievement/handler/AbstractEventHandler.java
@@ -2,8 +2,8 @@ package faang.school.achievement.handler;
 
 import faang.school.achievement.model.Achievement;
 import faang.school.achievement.model.AchievementProgress;
-import faang.school.achievement.service.achievement.AchievementCache;
-import faang.school.achievement.service.achievement.AchievementService;
+import faang.school.achievement.service.AchievementCache;
+import faang.school.achievement.service.AchievementService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.scheduling.annotation.Async;
