+++ src/main/java/faang/school/achievement/handler/impl/album/LibrarianAchievementHandler.java
@@ -2,8 +2,8 @@ package faang.school.achievement.handler.impl.album;
 
 import faang.school.achievement.event.album.AlbumCreatedEvent;
 import faang.school.achievement.handler.AbstractEventHandler;
-import faang.school.achievement.service.achievement.AchievementCache;
-import faang.school.achievement.service.achievement.AchievementService;
+import faang.school.achievement.service.AchievementCache;
+import faang.school.achievement.service.AchievementService;
 import org.springframework.stereotype.Service;
 
 @Service
