+++ src/main/java/faang/school/achievement/handler/impl/project/BusinessmanAchievementHandler.java
@@ -2,8 +2,8 @@ package faang.school.achievement.handler.impl.project;
 
 import faang.school.achievement.event.project.ProjectEvent;
 import faang.school.achievement.handler.AbstractEventHandler;
-import faang.school.achievement.service.achievement.AchievementCache;
-import faang.school.achievement.service.achievement.AchievementService;
+import faang.school.achievement.service.AchievementCache;
+import faang.school.achievement.service.AchievementService;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Component;
