+++ src/main/java/faang/school/achievement/service/AchievementCache.java
@@ -1,7 +1,5 @@
-package faang.school.achievement.service.achievement;
+package faang.school.achievement.service;
 
-import faang.school.achievement.dto.AchievementDto;
-import faang.school.achievement.mapper.achievement.AchievementMapper;
 import faang.school.achievement.model.Achievement;
 import faang.school.achievement.repository.AchievementRepository;
 import jakarta.annotation.PostConstruct;
