+++ src/main/resources/db/changelog/db.changelog-master.yaml
@@ -3,7 +3,7 @@ databaseChangeLog:
       file: db/changelog/changeset/achievement_V001__initial.sql
   - include:
       file: db/changelog/changeset/achievement_setup_V002.sql
-#  - include:
-#      file: db/changelog/changeset/achievement_setup_V003.sql
+  - include:
+      file: db/changelog/changeset/achievement_setup_V003.sql
   - include:
       file: db/changelog/changeset/achievement_setup_V004.sql
